requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.47.0
supabase==2.7.4
pandas==2.2.3
openpyxl==3.1.5
pydantic==2.9.2
httpx==0.27.2
